create new public classes NetheriteScaleMetrics and NetheriteScaleMonitor, add xml docs
